Move `mf_dedent` to `string_helpers.py`.
