+++ metricflow-semantics/tests_metricflow_semantics/experimental/collection_helpers/test_singleton.py
@@ -4,8 +4,7 @@ import logging
 
 import pytest
 from metricflow_semantics.experimental.test_helpers.performance_helpers import assert_performance_factor
-from metricflow_semantics.formatting.formatting_helpers import mf_dedent
-from metricflow_semantics.helpers.string_helpers import mf_newline_join
+from metricflow_semantics.helpers.string_helpers import mf_newline_join, mf_dedent
 
 from tests_metricflow_semantics.experimental.collection_helpers.singleton_test_classes import (
     PATH_TO_SINGLETON_TEST_CLASS_PY_FILE,
