+++ metricflow-semantics/tests_metricflow_semantics/mf_logging/test_pformat_dict.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import logging
 import textwrap
 
-from metricflow_semantics.formatting.formatting_helpers import mf_dedent
+from metricflow_semantics.helpers.string_helpers import mf_dedent
 from metricflow_semantics.mf_logging.pretty_print import mf_pformat_dict
 
 logger = logging.getLogger(__name__)
